move clicklistners to child classes
